Repository for my hobby OS dev project.
The short term goal is to get a basic protected mode kernel up and running,
first a hello world version (print a static message in protected mode), and
then an echo program, which echoes keyboard input on the screen. I will be
using GRUB as a bootloader, but I plan to implement multiboot compliance,
so this should conceivably be able to boot from any multiboot loader.

In the long run, it would be interesting to experiment with some micro
kernel concepts. But that is a long way off yet.