Implemented code to initialise GDT and IDT structures.
Also modified start.s; the lgdt and lidt assembly instructions must be
given a structure that contains a pointer to the table (GDT or IDT) and
also the size of that table. To implement this, I have defined two structs
for each table, the table itself, and also a descriptor for the table.