Added two header files that will define keycodes and mappings from scancodes to the corresponding keycode.
Scancodes are the code generated by the keyboard itself, but they can be
multiple bytes long, which is not ideal for fast O(1) lookups. Instead,
the scancode is converted into a keycode, which is itself used as an
index into an array of handler functions for each key.

I have also made a separate branch to develop the keyboard driver, since
the build of the hello world kernel still needs to be completed. I have
not yet added the top level makefile to the repository.