Created a source file for the keyboard driver. Since it will be a kernel mode driver, it will be compiled into the kernel.
keyboard.c will contain code for setting up the tables to map scancodes
to keycodes, and setting up keypress handlers. Also, the interrupt service
routine and the code that does the actual lookups will be in this file.

Some of the code for initialising the PS/2 controller will be in another
file.

So far, I have implemented a few of the keypress handlers, specifically
for managing the states of SHIFT, CTRL, caps lock, num lock etc.