More work on the keyboard module.
Substantial changes to scancodes.h, since I had mistyped the static array
initialisation.

Also many changes to keyboard.c, including a function to set keyboard LEDs.
The rest is minor changes.

I still need to add a function to initialise the keyboard hardware, which
needs to be called from the main kernel init function, and also an
interrupt handler, and code to set it up.