Added an interrupt handler for the keyboard.
I have also modified the functions that interpret a scancode such that they
read the scancode directly from the keyboard one byte at a time, rather
than having the entire scancode in a memory buffer. This way, we do not
need to know how long a scancode is in advance.

Also made the clear_buffer function in ps2.c public, so that it can be
used in the keyboard module. If there was a really long scancode, or one
that was not in our map, the hardware buffer will be cleared to move on to
the next key event.