Added top level makefile, and a script that automates the process of preparing a disk image file.
Note that due to the use of loopback devices and the mount command, the
script must be run as root.

make clean will unmount the disk image and disconnect the loop devices;
make scrub will remove the disk image file and the directory that the
makefile mounts it to. As well as recursing through subdirectories of
course.