Implemented a more sophisticated way of managing dependencies in the makefile.
Previously, dependencies were kept in a single file, which would have to be
rebuilt every time any source file was changed. Now, dependencies are kept
in separate files for each source file, so that when a source file is
modified, we only need to rebuild the dependencies for that source file.

All the dependency files are kept in a subdirectory called .depend (hidden
on *nix systems) which is not added to version control, of course. Make
will create the subdirectory if it does not exist (there is a rule for this
as part of the "all" target).

I have also removed the list of C source files in the makefile, since this
list was only needed by the old depend target. Now we only need the list of
object files that need to be built, which is nice: no more duplicated file
names :)