New code in interrupts.c that configures service routines in the interrupt descriptor table for all of the hardware interrupts.
Changed some of the preprocessor stuff in descriptors.h; instead of a
macro for the type field of an IDT entry, I have got constants for a few
possible type values. I have not included constants for 16 bit interrupt
descriptors, since I do not intend to use that at this stage.

Also added a header file with dummy C declarations for the hardware
interrupt service routines that are implemented in irqs.s.

As of this commit, the kernel builds with no errors, and I believe that
everything is there; now for testing.